refactor: pull out into queries folder
