+++ src/components/Selection/Dropdown.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import gql from 'gql-tag';
   import { onMount, createEventDispatcher } from 'svelte';
   import type { Generation1, PokemonGQL, PokemonLookupGQL } from 'src/types';
+  import first20Gen1Query from 'queries/first20Gen1Query';
   import { createLookupByNameGql, capitalizeFirstLetter } from 'utils';
   import { fetchGraphQL } from 'utils/network';
 
@@ -9,25 +9,6 @@
 
   let pokemonLookup: PokemonLookupGQL = {};
   let pokemonEntries: PokemonGQL[] = [];
-
-  const first20Gen1Query = gql`
-    query First20Gen1($id: Int!) {
-      gen1: pokemon_v2_generation_by_pk(id: $id) {
-        name
-        species: pokemon_v2_pokemonspecies(limit: 20, order_by: { id: asc }) {
-          pokemon: pokemon_v2_pokemons(limit: 1) {
-            id
-            name
-            forms: pokemon_v2_pokemonforms {
-              sprites: pokemon_v2_pokemonformsprites {
-                sprites
-              }
-            }
-          }
-        }
-      }
-    }
-  `;
 
   onMount(async () => {
     const result = await fetchGraphQL<{ data: Generation1 }>(first20Gen1Query, { id: 1 }, 'First20Gen1');
